+++ changelog.md
@@ -12,6 +12,8 @@
 * [added] Suggest code extraction from the name of identity 
 * [changed] Priority of loading media files in the detail view
 * [added] Orientation detection introduced
+* [added] Filter of Localitites and Identities
+* [added] Area added to Locality
 
 ## 01-2025
 * [added] Added carousel for media files in the detail view